Move gameplay logic from graphics into world
This seperated the game logic and the graphics engine even better.

The camera is game logic specific and the graphics engine only needs to know what kind of projection and orthographies matrix is used.
