+++ korangar/src/world/cameras/point_shadow.rs
