+++ korangar/src/world/cameras/start.rs
@@ -2,7 +2,8 @@ use std::f32::consts::{FRAC_PI_2, FRAC_PI_4};
 
 use cgmath::{Matrix4, Point3, Rad, SquareMatrix, Vector2, Vector3};
 
-use super::{perspective_reverse_lh, Camera};
+use super::Camera;
+use crate::graphics::perspective_reverse_lh;
 
 const DEFAULT_ZOOM: f32 = 150.0;
 const ROTATION_SPEED: f32 = 0.03;
